started comments Presenter
